refactor: move committee models into submodule
